Please specify a 'version' field in the manifest
It should match the format <upstreamversion>~ynh<packageversion>. For example : 4.3-2~ynh3. It is composed of the upstream version number (in the example, 4.3-2) and an incremental number for each change in the package without upstream change (in the example, 3). This incremental number can be reset to 1 each time the upstream version changes.